feat(frontend): replace AuthContext with Zustand auth store (P2-2)
- Add src/stores/authStore.js (Zustand)
- Remove src/contexts/AuthContext.jsx
- Update App.jsx: remove AuthProvider wrapper
- Update Login.jsx, Header.jsx, Sidebar.jsx: switch import to authStore
- Navigation (login→/ , logout→/login) moved to component layer
