+++ frontend/package.json
@@ -20,7 +20,8 @@
     "react-router-dom": "^7.13.1",
     "shadcn": "^4.0.8",
     "tailwind-merge": "^3.5.0",
-    "tw-animate-css": "^1.4.0"
+    "tw-animate-css": "^1.4.0",
+    "zustand": "^5.0.12"
   },
   "devDependencies": {
     "@eslint/js": "^9.39.4",
